sso/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cryptography==3.3.2
Django==3.1.9
djangorestframework==3.11.2
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
importlib-metadata==2.0.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.5
zipp==3.2.0
